functional-tests/pyproject.toml
[tool.poetry]
name = "proker_functional_tests"
version = "0.1.0"
description = ""
authors = ["Tom Gurion <nagasaki45@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
pytest = "^7.1.2"
pytest-playwright = "^0.3.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
